chore: move two comment components into one for ci
