+++ packages/components/src/CommentList/CommentList.tsx
@@ -2,28 +2,142 @@ import { useEffect, useState } from 'react'
 import { Box, Flex } from 'theme-ui'
 
 import { Button } from '../Button/Button'
-import { CommentContainer } from '../CommentContainer/CommentContainer'
+import { ButtonShowReplies } from '../ButtonShowReplies/ButtonShowReplies'
+import { CommentItem } from '../CommentItem/CommentItem'
+import { CreateReply } from '../CreateReply/CreateReply'
+import { Icon } from '../Icon/Icon'
 
 import type { IComment } from '../CommentItem/types'
 
 const MAX_COMMENTS = 5
 
-export interface IProps {
-  supportReplies?: boolean
-  comments: IComment[]
+interface IPropsShared {
   handleDelete: (_id: string) => Promise<void>
   handleEdit: (_id: string, comment: string) => Promise<void>
   handleEditRequest: () => Promise<void>
-  highlightedCommentId?: string
   isLoggedIn: boolean
-  maxLength: number
   isReplies: boolean
-  onMoreComments?: () => void
+  maxLength: number
   onSubmitReply?: (_id: string, reply: string) => Promise<void>
+}
+
+export interface IPropsCommentContainer extends IPropsShared {
+  comment: IComment
+  handleCommentReply?: (commentId: string | null) => void
+  supportReplies: boolean
+}
+
+export interface IPropsCommentList extends IPropsShared {
+  supportReplies?: boolean
+  comments: IComment[]
+  highlightedCommentId?: string
+  onMoreComments?: () => void
   setCommentBeingRepliedTo?: (commentId: string | null) => void
 }
 
-export const CommentList = (props: IProps) => {
+export const CommentContainer = (props: IPropsCommentContainer) => {
+  const [isShowReplies, setIsShowReplies] = useState(false)
+  const {
+    comment,
+    handleDelete,
+    handleEditRequest,
+    handleEdit,
+    supportReplies,
+    isLoggedIn,
+    isReplies,
+    maxLength,
+    onSubmitReply,
+  } = props
+  const { _id, creatorName, replies } = comment
+
+  const replyArrow = () => {
+    return (
+      <Box sx={{ paddingTop: 4 }}>
+        <Icon glyph="arrow-curved-bottom-right" />
+      </Box>
+    )
+  }
+
+  const repliesButton = () => {
+    return (
+      <ButtonShowReplies
+        creatorName={creatorName}
+        isShowReplies={isShowReplies}
+        replies={replies || []}
+        setIsShowReplies={() => setIsShowReplies(!isShowReplies)}
+      />
+    )
+  }
+
+  const createReply = () => {
+    if (onSubmitReply) {
+      return (
+        <CreateReply
+          commentId={_id}
+          isLoggedIn={isLoggedIn}
+          maxLength={maxLength}
+          onSubmit={onSubmitReply}
+        />
+      )
+    }
+  }
+
+  return (
+    <Box
+      sx={{
+        backgroundColor: 'white',
+        borderRadius: 1,
+        padding: 3,
+      }}
+    >
+      <CommentItem
+        comment={comment}
+        handleEditRequest={handleEditRequest}
+        handleDelete={handleDelete}
+        handleEdit={handleEdit}
+        isReply={isReplies ? true : false}
+      />
+
+      {supportReplies && !isShowReplies && repliesButton()}
+
+      {supportReplies && isShowReplies ? (
+        <Flex
+          sx={{
+            alignItems: 'stretch',
+            flexDirection: 'row',
+          }}
+        >
+          {replies && replyArrow()}
+
+          <Flex
+            sx={{
+              alignItems: 'stretch',
+              flexDirection: 'column',
+              flex: 1,
+            }}
+          >
+            <CommentList
+              comments={replies || []}
+              handleDelete={handleDelete}
+              handleEdit={handleEdit}
+              handleEditRequest={handleEditRequest}
+              isLoggedIn={isLoggedIn}
+              isReplies={true}
+              maxLength={maxLength}
+              supportReplies={false}
+            />
+
+            {createReply()}
+
+            {repliesButton()}
+          </Flex>
+        </Flex>
+      ) : null}
+    </Box>
+  )
+}
+
+export const CommentList = (props: IPropsCommentList) => {
   const {
     comments,
     handleDelete,
