+++ packages/components/src/index.ts
@@ -2,7 +2,6 @@ export { ArticleCallToAction } from './ArticleCallToAction/ArticleCallToAction'
 export { Button } from './Button/Button'
 export { ButtonShowReplies } from './ButtonShowReplies/ButtonShowReplies'
 export { Category } from './Category/Category'
-export { CommentContainer } from './CommentContainer/CommentContainer'
 export { CommentItem } from './CommentItem/CommentItem'
 export { CommentList } from './CommentList/CommentList'
 export { CreateReply } from './CreateReply/CreateReply'
